# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API de Torneos",
        "description": "Esta es la documentación de la API para la gestión de torneos.",
        "contact": {
            "email": "soporte@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/tournaments/simulate": {
            "post": {
                "tags": [
                    "Tournaments"
                ],
                "summary": "Simular torneo",
                "description": "Simula un torneo con los datos proporcionados.",
                "operationId": "9c643a5142ca15d7206a21e1c642c8d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "gender": {
                                        "description": "Género del torneo",
                                        "type": "string",
                                        "enum": [
                                            "Masculino",
                                            "Femenino"
                                        ],
                                        "example": "Femenino"
                                    },
                                    "type": {
                                        "description": "Tipo de torneo (individual o dobles)",
                                        "type": "string",
                                        "enum": [
                                            "single",
                                            "doubles"
                                        ],
                                        "example": "doubles"
                                    },
                                    "players": {
                                        "description": "Cantidad de jugadores participantes (mínimo 8)",
                                        "type": "integer",
                                        "example": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Torneo simulado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "winner": {
                                            "description": "Ganador del torneo",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 21
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Serena Williams"
                                                    },
                                                    "skill_level": {
                                                        "type": "integer",
                                                        "example": 97
                                                    },
                                                    "gender": {
                                                        "type": "string",
                                                        "example": "Femenino"
                                                    },
                                                    "strength": {
                                                        "type": "integer",
                                                        "example": 0
                                                    },
                                                    "speed": {
                                                        "type": "integer",
                                                        "example": 0
                                                    },
                                                    "reaction_time": {
                                                        "type": "integer",
                                                        "example": 95
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-12-23T21:33:39.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-12-23T21:33:39.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "match_results": {
                                            "description": "Resultados de los partidos",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID del partido",
                                                        "type": "integer",
                                                        "example": 4
                                                    },
                                                    "team_a": {
                                                        "description": "Equipo A",
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 27
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "Petra Kvitová"
                                                                },
                                                                "skill_level": {
                                                                    "type": "integer",
                                                                    "example": 89
                                                                },
                                                                "gender": {
                                                                    "type": "string",
                                                                    "example": "Femenino"
                                                                },
                                                                "strength": {
                                                                    "type": "integer",
                                                                    "example": 0
                                                                },
                                                                "speed": {
                                                                    "type": "integer",
                                                                    "example": 0
                                                                },
                                                                "reaction_time": {
                                                                    "type": "integer",
                                                                    "example": 86
                                                                },
                                                                "created_at": {
                                                                    "type": "string",
                                                                    "format": "date-time",
                                                                    "example": "2024-12-23T21:33:39.000000Z"
                                                                },
                                                                "updated_at": {
                                                                    "type": "string",
                                                                    "format": "date-time",
                                                                    "example": "2024-12-23T21:33:39.000000Z"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "team_b": {
                                                        "description": "Equipo B",
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 21
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "Serena Williams"
                                                                },
                                                                "skill_level": {
                                                                    "type": "integer",
                                                                    "example": 97
                                                                },
                                                                "gender": {
                                                                    "type": "string",
                                                                    "example": "Femenino"
                                                                },
                                                                "strength": {
                                                                    "type": "integer",
                                                                    "example": 0
                                                                },
                                                                "speed": {
                                                                    "type": "integer",
                                                                    "example": 0
                                                                },
                                                                "reaction_time": {
                                                                    "type": "integer",
                                                                    "example": 95
                                                                },
                                                                "created_at": {
                                                                    "type": "string",
                                                                    "format": "date-time",
                                                                    "example": "2024-12-23T21:33:39.000000Z"
                                                                },
                                                                "updated_at": {
                                                                    "type": "string",
                                                                    "format": "date-time",
                                                                    "example": "2024-12-23T21:33:39.000000Z"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "winner": {
                                                        "description": "Equipo ganador",
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 21
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "Serena Williams"
                                                                },
                                                                "skill_level": {
                                                                    "type": "integer",
                                                                    "example": 97
                                                                },
                                                                "gender": {
                                                                    "type": "string",
                                                                    "example": "Femenino"
                                                                },
                                                                "reaction_time": {
                                                                    "type": "integer",
                                                                    "example": 95
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "score_a": {
                                                        "description": "Puntaje del Equipo A",
                                                        "type": "integer",
                                                        "example": 181
                                                    },
                                                    "score_b": {
                                                        "description": "Puntaje del Equipo B",
                                                        "type": "integer",
                                                        "example": 190
                                                    },
                                                    "round": {
                                                        "description": "Número de ronda",
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Datos inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocurrió un error al simular el torneo."
                                        },
                                        "details": {
                                            "type": "string",
                                            "example": "Error inesperado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}